Bridge-API/requirements.txt
aniso8601==9.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
DateTime==4.3
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
protobuf==3.16.0
PyJWT==2.0.1
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.9
urllib3==1.26.4
uuid==1.30
Werkzeug==1.0.1
zope.interface==5.4.0